pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rfai"
version = "0.1.0"
description = "Recursive Fractal AI reference implementation"
authors = [{name = "Autogenerated", email = "auto@example.com"}]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=1.25",
]

[project.optional-dependencies]
extra = ["cupy-cuda11x>=11.0; platform_system == 'Linux'"]

[project.urls]
Homepage = "https://example.com/rfai"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.setuptools.packages.find]
where = ["."]
include = ["rfai*"]

